Finished debugging particle filter
I modified num_particles to 100, so there is more
particles for the particle filter. I noticed the
first problem appeared in updateWeights() function.
When I was accessing sensor observations, I previously
used the wrong index. I used i instead of j. So,
it was not guaranteed that the code accessed
observations vector within index range. The second
problem appeared also in updateWeights(), but
occurred when I calculated the weight of each
particle. I did not do a check for verifying that
the observations landmark_id existed in the
predicted_landmarks vector. So, initially, I was
getting weights equal to 0, but that went away after
I added a loop to go through all predicted_landmarks
and check that observations id existed in
predicted_landmarks. Thus, I no longer got garbage
data for mu_x and mu_y and they could be correctly
applied to the multivarian gaussian equation. For
resample, I made sure to generate a random index
within the range of particles by multiplying
 with num_particles-1 instead of num_particles.
I also did a check to verify index was within range
of particles accessible.